backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
torch==2.0.1
langdetect==1.0.9
googletrans==4.0.0rc1
deep-translator==1.11.4
ddgs==0.9.9
numpy==1.24.3
urllib3==2.0.4
lxml==4.9.3
groq==0.4.1
feedparser==6.0.10